# server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/user":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}}}},"/user/{id}":{"get":{"operationId":"UserController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"UserController_update","summary":"Create a new location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"operationId":"UserController_remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/location":{"get":{"operationId":"LocationController_findAll","summary":"Get all locations","parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Location"}}}}}},"tags":["Location"]},"post":{"operationId":"LocationController_create","summary":"Create a new location","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]}},"/location/{id}":{"get":{"operationId":"LocationController_findById","summary":"Get a location by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]},"put":{"operationId":"LocationController_update","summary":"Create a new location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]},"delete":{"operationId":"LocationController_remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Location"}}}}},"tags":["Location"]}},"/store":{"get":{"operationId":"StoreController_findAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"StoreController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreDto"}}}},"responses":{"201":{"description":""}}}},"/store/{id}":{"get":{"operationId":"StoreController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"StoreController_update","summary":"Create a new location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Store"}}}}}},"delete":{"operationId":"StoreController_remove","summary":"Delete a location","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Store"}}}}}}}},"info":{"title":"Ludu API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserDto":{"type":"object","properties":{"username":{"type":"string","example":"popolito","description":"Your nickname"}},"required":["username"]},"User":{"type":"object","properties":{}},"Location":{"type":"object","properties":{}},"LocationDto":{"type":"object","properties":{"name":{"type":"string","example":"Location 1","description":"The name of the location"},"stores":{"description":"Stores","type":"array","items":{"type":"string"}}},"required":["name"]},"StoreDto":{"type":"object","properties":{"name":{"type":"string","example":"Location 1","description":"The name of the store"},"address":{"type":"string","example":"75 rue de test","description":"The address of the store"},"owner":{"type":"string"},"phone":{"type":"string"},"location":{"type":"string"}},"required":["name","address","owner","phone","location"]},"Store":{"type":"object","properties":{}}}}}